refactor(api): centralize performance telemetry types
